requirements.txt
#azure-storage-queue==2.1.0
azure-storage-queue